SERVER/requirements.txt
Flask-Dance
PyJWT
flask-socketio
eventlet
Flask>=2.0
Flask-SQLAlchemy>=2.5
Werkzeug>=2.0
python-dotenv>=1.0
psycopg2-binary
flask_cors